refactor: enhance module resolution and clean up context handling by removing unused files
