--- libs/orpc/src/context.test.ts
@@ -1,11 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { createContext } from "./context";
-
-describe("createContext", () => {
-  it("should create a context", async () => {
-    const context = await createContext({
-      req: new Request("http://localhost"),
-    });
-    expect(context).toBeDefined();
-  });
-});
