+++ libs/orpc/src/handlers/helius/getTokenList.handler.ts
@@ -1,5 +1,5 @@
+import type { GetTokenListInput } from "@/schemas/helius/getTokenList.schema";
 import { helius } from "../../helius";
-import type { GetTokenListInput } from "../../schemas/helius/getTokenList.schema";
 
 export async function getTokenListHandler({ input }: GetTokenListInput) {
   return await helius.rpc.searchAssets({
