+++ libs/orpc/src/index.ts
@@ -1,4 +1,3 @@
 export { client } from "./client";
-export { type Context, createContext } from "./context";
 export { appRouter } from "./routers/app.router";
 export { rpcHandler } from "./rpcHandler";
