+++ libs/orpc/src/procedures/helius/getTokenList.ts
@@ -1,6 +1,6 @@
-import { getTokenListHandler } from "../../handlers/helius/getTokenList.handler";
-import { getTokenListInputSchema } from "../../schemas/helius/getTokenList.schema";
-import { baseProcedure } from "../baseProcedure";
+import { getTokenListHandler } from "@/handlers/helius/getTokenList.handler";
+import { baseProcedure } from "@/procedures/baseProcedure";
+import { getTokenListInputSchema } from "@/schemas/helius/getTokenList.schema";
 
 export const getTokenList = baseProcedure
   .input(getTokenListInputSchema)
