+++ libs/orpc/src/rpcHandler.ts
@@ -1,6 +1,6 @@
 import { RPCHandler } from "@orpc/server/fetch";
 import { CORSPlugin } from "@orpc/server/plugins";
-import { appRouter } from "./routers/app.router";
+import { appRouter } from "@/routers/app.router";
 
 export const rpcHandler = new RPCHandler(appRouter, {
   plugins: [
